requirements.txt
streamlit
pandas
joblib
scikit-learn
pyyaml
